add som args

+++ cmds/say.js
@@ -1,4 +1,5 @@
 exports.run = (bot, msg, args) => {
+    if(!args) return msg.channel.send(":x: | Please Input An Arguments To Say!")
     msg.channel.send(args.join(" "));
     msg.delete()
 }
